refactor: Create shared space and refactor
